// src/app/interfaces/employer-audit.ts
export interface EmployerAudit {
  case_number: any;
  employer_name: any;
  employer_business_dba: any;
  employer_address: any;
  employer_city: any;
  employer_state: any;
  employer_postal_code: any;
  employer_country: any;
  employer_province: any;
  employer_phone: any;
  employer_phone_ext: any;
  date_changed: any;
}